Restructure dashboard files
