+++ terminal-backend/src/main/java/org/fyp24064/dashboard/model/StockData.java
@@ -1,7 +1,6 @@
-package org.fyp24064;
+package org.fyp24064.dashboard.model;
 
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
@@ -12,20 +11,9 @@ public class StockData {
     private String symbol;
     private List<HistoricalQuote> entries;
 
-    // method for future use
+    // method for future use: real time refresh
     public void addEntry(HistoricalQuote entry) {
         entries.add(entry);
-    }
-
-    @Getter
-    @Setter
-    public static class HistoricalQuote {
-        private Date date;
-        private Double open;
-        private Double high;
-        private Double low;
-        private Double close;
-        private Double volume;
     }
 
     public StockData() {
