+++ terminal-ui/pom.xml
@@ -59,12 +59,6 @@
         </dependency>
 
         <dependency>
-            <groupId>org.jfree</groupId>
-            <artifactId>jfreechart</artifactId>
-            <version>1.5.3</version>
-        </dependency>
-      
-        <dependency>
             <groupId>org.fyp24064</groupId>
             <artifactId>terminal-backend</artifactId>
             <version>1.0-SNAPSHOT</version>
@@ -80,25 +74,6 @@
             <groupId>com.panemu</groupId>
             <artifactId>tiwulfx-dock</artifactId>
             <version>0.5</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.httpcomponents.client5</groupId>
-            <artifactId>httpclient5-fluent</artifactId>
-            <version>5.4.1</version>
-        </dependency>
-
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-            <version>2.18.2</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.projectlombok</groupId>
-            <artifactId>lombok</artifactId>
-            <version>1.18.36</version>
-            <scope>provided</scope>
         </dependency>
 
     </dependencies>
