+++ terminal-ui/src/main/java/org/fyp24064/NewsDisplay.java
@@ -5,6 +5,8 @@ import java.util.List;
 import javafx.scene.control.Label;
 import javafx.scene.control.ScrollPane;
 import javafx.scene.layout.VBox;
+import org.fyp24064.dashboard.model.StockNewsResult;
+import org.fyp24064.dashboard.service.StockService;
 
 public class NewsDisplay {
     private StockService stockService;
@@ -18,12 +20,12 @@ public class NewsDisplay {
         newsEntries.getStyleClass().add("pane");
 
         try {
-            List<StockNews.Result> newsResults = stockService.getNews(stockSymbol);
+            List<StockNewsResult> newsResults = stockService.getNews(stockSymbol);
             if (newsResults == null || newsResults.isEmpty()) {
                 return null;
             }
             int newsCount = 1;
-            for (StockNews.Result newsEntry : newsResults) {
+            for (StockNewsResult newsEntry : newsResults) {
                 newsEntries.getChildren().add(createNewsEntry(newsEntry, stockSymbol, newsCount));
                 newsCount++;
             }
@@ -38,7 +40,7 @@ public class NewsDisplay {
         return root;
     }
 
-    private VBox createNewsEntry(StockNews.Result newsEntry, String ticker, int newsCount) {
+    private VBox createNewsEntry(StockNewsResult newsEntry, String ticker, int newsCount) {
         Label content = new Label(
                 newsCount + ". \""
                 + newsEntry.getTitle() + "\"\n"
